+++ R/single_simulation_landscape.R
@@ -1,13 +1,13 @@
-#' Make 2D density-based landscape plot for a single simulation output
+#' Make 2D static landscape plot for a single simulation output
 #'
 #' @param output A matrix of simulation output.
 #' @param x The name of the target variable.
-#' @param adjust,from,to Passed to \code{density}.
+#' @param adjust,from,to Passed to `density`.
 #' @param Umax The maximum displayed value of potential.
 #'
-#' @return A \code{2d_density_landscape} object that describes the landscape of the system, including the smooth distribution and the landscape plot.
+#' @return A `2d_static_landscape` object that describes the landscape of the system, including the smooth distribution and the landscape plot.
 #' @export
-make_2d_density <- function(output, x, adjust = 50, from = -0.1, to = 1, Umax = 5) {
+make_2d_static <- function(output, x, adjust = 50, from = -0.1, to = 1, Umax = 5) {
   d <- stats::density(output[, x], adjust = adjust, from = from, to = to)
   p <- data.frame(x = d$x, y = d$y, U = pmin(-log(d$y), Umax)) %>%
     ggplot2::ggplot(mapping = ggplot2::aes(x = x, y = U)) +
@@ -17,17 +17,37 @@ make_2d_density <- function(output, x, adjust = 50, from = -0.1, to = 1, Umax = 
     ggplot2::xlab(x)
 
   result <- list(dist = d, plot = p, x = x, adjust = adjust, from = from, to = to, Umax = Umax)
-  class(result) <- c("2d_density_landscape", "landscape")
+  class(result) <- c("2d_static_landscape", "landscape")
   return(result)
 }
 
-#' Make a tidy \code{data.frame} from smooth 2d distribution matrix
+
+#' Make 2D density-based landscape plot for a single simulation output
+#' @description
+#' `r lifecycle::badge("deprecated")`
 #'
-#' @param dist_2d \code{kde2d} distribution.
+#' This function was deprecated. Use [make_2d_static()] instead.
+#'
+#' @param output A matrix of simulation output.
+#' @param x The name of the target variable.
+#' @param adjust,from,to Passed to `density`.
+#' @param Umax The maximum displayed value of potential.
+#'
+#' @return A `2d_static_landscape` object that describes the landscape of the system, including the smooth distribution and the landscape plot.
+#' @export
+#' @export
+make_2d_density <- function(output, x, adjust = 50, from = -0.1, to = 1, Umax = 5) {
+  lifecycle::deprecate_warn("0.1.3", "make_2d_density()", "make_2d_static()")
+  make_2d_static(output, x, adjust, from, to, Umax)
+}
+
+#' Make a tidy `data.frame` from smooth 2d distribution matrix
+#'
+#' @param dist_2d `kde2d` distribution.
 #' @param value The value of the variable of interest.
 #' @param var_name The name of the variable.
 #'
-#' @return A tidy \code{data.frame}.
+#' @return A tidy `data.frame`.
 #'
 #' @export
 #'
@@ -53,7 +73,7 @@ make_2d_tidy_dist <- function(dist_2d, value = NULL, var_name = NULL) {
 #'
 #' @param base The base of logarithm
 #'
-#' @return A \code{trans} scale object from the \code{scales} package.
+#' @return A `trans` scale object from the `scales` package.
 #'
 #' @export
 reverselog_trans <- function(base = exp(1)) {
@@ -70,13 +90,13 @@ reverselog_trans <- function(base = exp(1)) {
 #'
 #' @param output A matrix of simulation output.
 #' @param x,y The name of the target variable.
-#' @param n,lims,h Passed to \code{\link[ks]{kde}} or \code{\link[MASS]{kde2d}}.
-#' If using \code{ks::kde}, \code{H = diag(h, 2, 2)}.
+#' @param n,lims,h Passed to [ks::kde()] or [MASS::kde2d()].
+#' If using `ks::kde`, `H = diag(h, 2, 2)`.
 #' Note: the definition of bandwidth (`h`) is different in two functions.
-#' To get a similar output, the `h` is about 50 to 5000 times smaller for \code{\link[ks]{kde}} than \code{\link[MASS]{kde2d}}
-#' @param kde_fun Which to use? Choices: "ks" \code{ks::kde} (default; faster and taking less memory); "MASS" \code{MASS::kde2d}.
+#' To get a similar output, the `h` is about 50 to 5000 times smaller for [ks::kde()] than [MASS::kde2d()]
+#' @param kde_fun Which to use? Choices: "ks" `ks::kde` (default; faster and taking less memory); "MASS" `MASS::kde2d`.
 #'
-#' @return A \code{kde2d}-type list of smooth distribution.
+#' @return A `kde2d`-type list of smooth distribution.
 #' @export
 make_2d_kernel_dist <- function(output, x, y, n = 200, lims = c(-0.1, 1.1, -0.1, 1.1), h, kde_fun = "ks") {
   if (is.list(output)) output <- output[[1]]
@@ -102,8 +122,7 @@ make_2d_kernel_dist <- function(output, x, y, n = 200, lims = c(-0.1, 1.1, -0.1,
     # reformat the result to the format of MASS::kde2d
     result <- list(x = result$eval.points[[1]], y = result$eval.points[[2]], z = pmax(result$estimate, 0)) # different result??
     return(result)
-  }
-  else {
+  } else {
     stop('Wrong input for `kde_fun`. Please choose from "MASS" and "ks".')
   }
 }
@@ -114,9 +133,9 @@ make_2d_kernel_dist <- function(output, x, y, n = 200, lims = c(-0.1, 1.1, -0.1,
 #' @param output A matrix of simulation output.
 #' @param x,y The name of the target variable.
 #' @param Umax The maximum displayed value of potential.
-#' @param n,lims,h,kde_fun Passed to \code{\link{make_2d_kernel_dist}}
+#' @param n,lims,h,kde_fun Passed to [make_2d_kernel_dist()]
 #'
-#' @return A \code{3d_static_landscape} object that describes the landscape of the system, including the smooth distribution and the landscape plot.
+#' @return A `3d_static_landscape` object that describes the landscape of the system, including the smooth distribution and the landscape plot.
 #'
 #' @export
 make_3d_static <- function(output, x, y, Umax = 5, n = 200, lims = c(-0.1, 1.1, -0.1, 1.1), h = 1e-3, kde_fun = "ks") {
@@ -128,14 +147,15 @@ make_3d_static <- function(output, x, y, Umax = 5, n = 200, lims = c(-0.1, 1.1, 
 
   message("Making the plot...")
   p <- plotly::plot_ly(x = out_2d$x, y = out_2d$y, z = pmin(-log(out_2d$z %>% t()), Umax), type = "surface")
-  p <- plotly::layout(p, scene = list(xaxis = list(title = x), yaxis = list(title = y), zaxis = list(title = "U")))
+  p <- plotly::layout(p, scene = list(xaxis = list(title = x), yaxis = list(title = y), zaxis = list(title = "U"))) %>% plotly::colorbar(title = "U")
   message("Done!")
 
   message("Making the 2d plot...")
   p2 <- ggplot2::ggplot(make_2d_tidy_dist(out_2d), ggplot2::aes(x = x, y = y)) +
     ggplot2::geom_raster(ggplot2::aes(fill = pmin(-log(z), Umax))) +
     ggplot2::scale_fill_viridis_c() +
-    ggplot2::labs(x = x, y = y, fill = "U")
+    ggplot2::labs(x = x, y = y, fill = "U") +
+    ggplot2::theme_bw()
   message("Done!")
 
   result <- list(dist = out_2d, plot = p, plot_2 = p2, x = x, y = y, Umax = Umax, n = n, lims = lims, h = h, kde_fun = kde_fun)
@@ -148,10 +168,10 @@ make_3d_static <- function(output, x, y, Umax = 5, n = 200, lims = c(-0.1, 1.1, 
 #'
 #' @param output A matrix of simulation output.
 #' @param x,y,z The name of the target variable.
-#' @param n,lims,h Passed to \code{\link[ks]{kde}} (but using the format of \code{\link[MASS]{kde2d}} to make it consistent across functions).
-#' For \code{ks::kde}, \code{H = diag(h, 2, 2)}.
+#' @param n,lims,h Passed to [ks::kde()] (but using the format of [MASS::kde2d()] to make it consistent across functions).
+#' For `ks::kde`, `H = diag(h, 2, 2)`.
 #'
-#' @return A \code{MASS::kde2d}-type list of smooth distribution.
+#' @return A `MASS::kde2d`-type list of smooth distribution.
 #' @export
 make_3d_kernel_dist <- function(output, x, y, z, n = 200, lims = c(-0.1, 1.1, -0.1, 1.1, -0.1, 1.1), h) {
   if (is.list(output)) output <- output[[1]]
@@ -176,13 +196,13 @@ make_3d_kernel_dist <- function(output, x, y, z, n = 200, lims = c(-0.1, 1.1, -0
   return(result)
 }
 
-#' Make a tidy \code{data.frame} from smooth 3d distribution matrix
+#' Make a tidy `data.frame` from smooth 3d distribution matrix
 #'
-#' @param dist_3d \code{kde2d}-type distribution.
+#' @param dist_3d `kde2d`-type distribution.
 #' @param value The value of the variable of interest.
 #' @param var_name The name of the variable.
 #'
-#' @return A tidy \code{data.frame}.
+#' @return A tidy `data.frame`.
 #'
 #' @export
 #'
@@ -211,9 +231,9 @@ make_3d_tidy_dist <- function(dist_3d, value = NULL, var_name = NULL) {
 #' @param output A matrix of simulation output.
 #' @param x,y,z The name of the target variable.
 #' @param Umax The maximum displayed value of potential.
-#' @param n,lims,h Passed to \code{\link{make_3d_kernel_dist}}
+#' @param n,lims,h Passed to [make_3d_kernel_dist()]
 #'
-#' @return A \code{4d_static_landscape} object that describes the landscape of the system, including the smoothed distribution and the landscape plot.
+#' @return A `4d_static_landscape` object that describes the landscape of the system, including the smoothed distribution and the landscape plot.
 #'
 #' @export
 make_4d_static <- function(output, x, y, z, Umax = 5, n = 50, lims = c(-0.1, 1.1, -0.1, 1.1, -0.1, 1.1), h = 1e-3) {
@@ -231,7 +251,8 @@ make_4d_static <- function(output, x, y, z, Umax = 5, n = 50, lims = c(-0.1, 1.1
     dplyr::filter(-log(.$d) < Umax) %>%
     plotly::plot_ly(x = ~x, y = ~y, z = ~z, color = -log(.$d)) %>%
     plotly::add_markers(size = I(5)) %>%
-    plotly::layout(scene = list(xaxis = list(title = x), yaxis = list(title = y), zaxis = list(title = z)))
+    plotly::layout(scene = list(xaxis = list(title = x), yaxis = list(title = y), zaxis = list(title = z))) %>%
+    plotly::colorbar(title = "U")
   message("Done!")
 
   result <- list(dist = df_tidy, plot = p, x = x, y = y, z = z, Umax = Umax, n = n, lims = lims, h = h)